Repository, Service, Controller 분할 가게 삭제 수정

+++ src/main/java/com/coffeemantang/ZMT_BACK/service/StoreService.java
@@ -43,48 +43,16 @@ public class StoreService {
         return storeRepository.save(storeEntity);
     }
 
-    //메뉴 추가
-    public MenuEntity addMenu(final MenuEntity menuEntity) {
+    //가게 삭제
+    public void deleteStore(int memberId, String storeId) {
 
-        if (menuEntity == null || menuEntity.getStoreId() == null) {
-            log.warn("StoreService.addMenu() : menuEntity에 내용이 부족해요");
-            throw new RuntimeException("StoreService.addMenu() : menuEntity에 내용이 부족해요");
+        StoreEntity storeEntity = storeRepository.findByStoreId(storeId);
+
+        if (memberId != storeEntity.getMemberId()) {
+            log.warn("StoreService.addOption() : 로그인된 유저와 가게 소유자가 다릅니다.");
+            throw new RuntimeException("StoreService.addOption() : 로그인된 유저와 가게 소유자가 다릅니다.");
         }
 
-        return menuRepository.save(menuEntity);
-
-    }
-
-    //메뉴 번호 생성 메서드
-    public int createMenuNumber() {
-
-        //menuNumber 컬럼만 가져옴
-        List<Integer> list = menuRepository.selectAllMenuNumber();
-
-        //리스트가 비어있으면 1, 아니면 최대값 + 1
-        if (list.isEmpty()) {
-            return 1;
-        } else {
-            //menuNumber에서 최대값
-            int max = Collections.max(list);
-            return max + 1;
-        }
-
-    }
-
-    //옵션 추가
-    public OptionEntity addOption(final OptionEntity optionEntity) {
-
-        if (optionEntity == null || optionEntity.getMenuId() == 0) {
-            log.warn("StoreService.addOption() : optionEntity에 내용이 부족해요");
-            throw new RuntimeException("StoreService.addOption() : optionEntity에 내용이 부족해요");
-        }
-
-        return optionRepository.save(optionEntity);
-
-    }
-
-    public void deleteStore(String storeId) {
         storeRepository.deleteById(storeId);
     }
 }
